GuiSys and Gui Editor, "Choice" window: Revised and fixed code regarding the selected choice.
git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@373 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ Libs/GuiSys/Window.cpp
@@ -153,7 +153,7 @@ WindowT::WindowT(const WindowT& Window, bool Recursive)
         BackRenderMat=MatSys::Renderer->RegisterMaterial(m_Gui.m_MaterialMan.GetMaterial(BackRenderMatName));
     }
 
-    for (unsigned long i=0; i<4; i++)
+    for (unsigned int i=0; i<4; i++)
     {
         Rect       [i]=Window.Rect[i];
         BackColor  [i]=Window.BackColor[i];
